pyproject.toml
[project]
name = "tccw-autoresearch"
version = "0.1.0"
description = "Agnostic autonomous improvement engine"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov"]

[project.scripts]
autoresearch = "autoresearch.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/autoresearch"]

[tool.pytest.ini_options]
testpaths = ["tests"]
